PIL-1271 - Refactored tests to move to the integration test folder/structure, made ITs runnable through sbt commands
